feat: reduce the size of the token detail position component
