+++ dapps/reports/src/views/sections/TokenDetailItems/TokenDetailByPositionContainer.tsx
@@ -1,4 +1,4 @@
-import CardsTokenDetailByPosition from '@karpatkey-monorepo/reports/src/views/sections/TokenDetailItems/CardsTokenDetailByPosition'
+import CardList from '@karpatkey-monorepo/reports/src/views/sections/TokenDetailItems/TokenDetailByPositionItems/CardList'
 import Filter from '@karpatkey-monorepo/shared/components/Filter/Filter'
 import Form from '@karpatkey-monorepo/shared/components/Filter/Form'
 import PaperSection from '@karpatkey-monorepo/shared/components/PaperSection'
@@ -88,7 +88,7 @@ const TokenDetailByPositionContainer = (props: TokenDetailByPositionContainerPro
 
   return (
     <PaperSection subTitle="Token detail by position" filter={filter}>
-      <CardsTokenDetailByPosition tokenDetailByPosition={tokenDetailByPosition} />
+      <CardList tokenDetailByPosition={tokenDetailByPosition} />
     </PaperSection>
   )
 }
